factor out searchable pdf view
